Landing page / Improve multilingual support (#4924)
* Landing page / Improve multilingual support

Landing page is displayed by default taking into account the Accept-Language of the browser. Then translations are displayed depending on that language if available. A specific language can be forced using the `language` parameter but the user does not know in which languages the record is described by default.

The `language=all` parameter can now be set to return the record description in all languages used in the record. All element language is identified by using `xml:lang` attribute.

A language switcher allows to display one language or all translations.

The HTML page meta is improved by providing `link` to all translations page representing the same resource (as described in https://support.google.com/webmasters/answer/189077). Search engine will be aware that different pages describe the same resource but in different languages (to avoid to create duplicates in search results).

The sitemap now point to this page by default.

* Landing page / Improve multilingual support / ISO19115-3 support.

* Landing page / Improve multilingual support / DC and ISO19110 fixes.

* Landing page / Improve multilingual support / ISO19110 fixes.

* Landing page / Improve multilingual support / Select default language

Fix also multilingual editing on ISO19115-3.